[wal] Move wal to per-epoch store (#6508)
